requirements.txt
streamlit
Pillow
requests
cloudinary
open-interpreter==0.1.15